==> Scripts/2a_Generate_AR_forecasts.Rmd <==
---
title: "fDOM_forecast_generation"
author: "Dexter Howard"
date: "2024-06-28"
output: html_document
---


## Load packages and functions 

```{r}
library(tidyverse)

source('./Functions/AR_4cast_function.R')

```


## Get data 

target data 
```{r}
### Read in df 
targets_fdom <- read_csv("../Data/GeneratedData/Targets_fDOM_allReservoirs.csv")

### FCR water Q data
fcr_waterQ <- targets_fdom |>
  filter(site_id == "fcre")

###BVR water Q data
bvr_waterQ <- targets_fdom |>
  filter(site_id == "bvre")

### CCR water Q data
ccr_waterQ <- targets_fdom |>
  filter(site_id == "ccre")


```


NOAA forecasts 
```{r}
## FCR and BVR 
fcr_bvr_noaa_daily <- read.csv("../Data/GeneratedData/FCR_NOAA_stage2_dailyaverage_27sep20-19mar25.csv")

## CCR
ccr_noaa_daily <- read.csv("../Data/GeneratedData/CCR_NOAA_stage2_dailyaverage_27sep20-19mar25.csv")

```


Water Temp forecasts 
```{r}
## FCR 
fcr_flare <- read.csv("../Data/GeneratedData/FCR_FLARE_11nov22-19mar25.csv")

##BVR 
bvr_flare <- read.csv("../Data/GeneratedData/BVR_FLARE_8nov22-20mar25.csv")

##CCR 
ccr_flare <- read.csv("../Data/GeneratedData/CCR_FLARE_2jan23-20mar25.csv")

```

Set up directories to hold forecasts

```{r}
##uncomment if folders don't exist locally
# dir.create("../Data/Forecasts/AR_4casts")
# dir.create("../Data/Forecasts/AR_4casts/fcre")
# dir.create("../Data/Forecasts/AR_4casts/bvre")
# dir.create("../Data/Forecasts/AR_4casts/ccre")

```


## Run AR forecasts for all reservoirs

FCR
```{r}
##set up inputs to function
model_id <- "fDOM_AR_dwh"
var <- "fDOM_QSU_mean"
site <- "fcre"
forecast_depths <- 1.6
project_id <- "vera4cast"
calibration_start_date <- ymd("2023-03-13") 

targets_df <- fcr_waterQ
water_temp_4cast_data <- fcr_flare
noaa_4cast <- fcr_bvr_noaa_daily

n_members <- 31

output_folder <- paste0((str_sub(getwd(), end = -8)), "Data/Forecasts/AR_4casts/fcre/", model_id, "_", site, "_")


##Loop for 34 day horizon windows
forecast_date <- seq(ymd("2025-02-08"), ymd("2025-02-28"), by = "day")
forecast_horizon <- 34


for (j in 1:length(forecast_date)) {

  print(forecast_date[j])

generate_fDOM_forecast(forecast_date = forecast_date[j], forecast_horizon = forecast_horizon, n_members = n_members,
                       output_folder = output_folder, calibration_start_date = calibration_start_date,
                       model_id = model_id, targets_df = targets_df,
                       water_temp_4cast_data = water_temp_4cast_data, noaa_4cast = noaa_4cast, var = var,
                      site = site, forecast_depths = forecast_depths, project_id = project_id)
}

## days loop broke: 07feb2025
## Days w/ NAs: none


```


BVR
```{r}

##set up inputs to function
model_id <- "fDOM_AR_dwh"
var <- "fDOM_QSU_mean"
site <- "bvre"
forecast_depths <- 1.5
project_id <- "bvr_fdom_hindcast"
calibration_start_date <- ymd("2023-03-13") 

targets_df <- bvr_waterQ
water_temp_4cast_data <- bvr_flare
noaa_4cast <- fcr_bvr_noaa_daily

n_members <- 31

output_folder <- paste0((str_sub(getwd(), end = -8)), "Data/Forecasts/AR_4casts/bvre/", model_id, "_", site, "_")


#Loop for 34 day horizon windows
forecast_date <- seq(ymd("2024-03-01"), ymd("2025-02-28"), by = "day")
forecast_horizon <- 34


for (j in 1:length(forecast_date)) {

  print(forecast_date[j])

generate_fDOM_forecast(forecast_date = forecast_date[j], forecast_horizon = forecast_horizon, n_members = n_members,
                       output_folder = output_folder, calibration_start_date = calibration_start_date,
                       model_id = model_id, targets_df = targets_df,
                       water_temp_4cast_data = water_temp_4cast_data, noaa_4cast = noaa_4cast, var = var,
                      site = site, forecast_depths = forecast_depths, project_id = project_id)

}

## Days w/ NAs: 2apr - 4apr2024


```


CCR
```{r}

##set up inputs to function
model_id <- "fDOM_AR_dwh"
var <- "fDOM_QSU_mean"
site <- "ccre"
forecast_depths <- 1.5
project_id <- "ccr_fdom_hindcast"
calibration_start_date <- ymd("2023-03-13") 

targets_df <- ccr_waterQ
water_temp_4cast_data <- ccr_flare
noaa_4cast <- ccr_noaa_daily

n_members <- 31

output_folder <- paste0((str_sub(getwd(), end = -8)), "Data/Forecasts/AR_4casts/ccre/", model_id, "_", site, "_")


#Loop for 34 day horizon windows
forecast_date <- seq(ymd("2024-03-01"), ymd("2025-02-28"), by = "day")
forecast_horizon <- 34


for (j in 1:length(forecast_date)) {

  print(forecast_date[j])

generate_fDOM_forecast(forecast_date = forecast_date[j], forecast_horizon = forecast_horizon, n_members = n_members,
                       output_folder = output_folder, calibration_start_date = calibration_start_date,
                       model_id = model_id, targets_df = targets_df,
                       water_temp_4cast_data = water_temp_4cast_data, noaa_4cast = noaa_4cast, var = var,
                      site = site, forecast_depths = forecast_depths, project_id = project_id)

}

## Days w/ NAs: 29dec2024 - 7jan2025; 10sep2024 - 15sep2024; 1aug - 4aug2024

```
