==> Scripts/2d_Generate_NNETAR_forecasts.Rmd <==
---
title: "fDOM_NNETAR_forecast_generation"
author: "Dexter Howard"
date: "2024-08-23"
output: html_document
---


## Load packages and functions 

```{r}
library(tidyverse)

source('./Functions/NNETAR_4cast_function.R')

```


## Get data 

target data 
```{r}
### Read in df 
targets_fdom <- read_csv("../Data/GeneratedData/Targets_fDOM_allReservoirs.csv") |> 
    filter(datetime >= ymd("2023-03-13")) #filter to same starting point as AR and met regression


```



Set up directories to hold forecasts

```{r}
##uncomment lines below if folder don't exist
# dir.create("../Data/Forecasts/NNETAR_4casts")
# dir.create("../Data/Forecasts/NNETAR_4casts/allRes")

```


## Run NNETAR forecasts for all reservoir


```{r}

forecast_date <- seq(ymd("2024-03-01"), ymd("2025-02-28"), by = "day")

var <- "fDOM_QSU_mean"
targets_df <- targets_fdom
forecast_horizon <- 35

output_folder <- paste0((str_sub(getwd(), end = -8)), "Data/Forecasts/NNETAR_4casts/allRes/", "NNETAR_allRes_")


for (j in 1:length(forecast_date)) {
  
  print(forecast_date[j]) 
  
  fdom_nnetar(forecast_date = forecast_date[j], targets_df = targets_df, var = var, 
              forecast_horizon = forecast_horizon, output_folder = output_folder)
  
      } #end for loop


```
