==> Scripts/3_Compile_4cast_outputs.Rmd <==
---
title: "Analyze fDOM hindcasts"
author: "Dexter Howard"
date: "2024-09-01"
output: html_document
---


```{r, message = F}
library(tidyverse)
library(score4cast)

```



## get targets data

```{r}
### Read in df 
targets_fdom <- read.csv("../Data/GeneratedData/Targets_fDOM_allReservoirs.csv")

#rename to match prior code
allres_targets <- targets_fdom  


```

## Compile AR 4cast outputs  

```{r}
#FCR AR forecasts
#bind forecast outputs into one data frame 
fcr_AR_output_all <- list.files(path = "../Data/Forecasts/AR_4casts/fcre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#BVR AR forecasts
#bind forecast outputs into one data frame 
bvr_AR_output_all <- list.files(path = "../Data/Forecasts/AR_4casts/bvre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#CCR AR forecasts
#bind forecast outputs into one data frame 
ccr_AR_output_all <- list.files(path = "../Data/Forecasts/AR_4casts/ccre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#Compile AR forecasts and write csv
allres_AR <- rbind(fcr_AR_output_all, bvr_AR_output_all, ccr_AR_output_all)

write.csv(allres_AR, "../Data/Forecasts/AR_4casts/allReservoir_AR_4cast.csv", row.names = F)


```



### Compile NNETAR 4cast outputs  

All reservoir NNETAR forecasts
```{r}
#bind forecast outputs into one data frame 
allRes_NNETAR_output_all <- list.files(path = "../Data/Forecasts/NNETAR_4casts/allRes", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#write csv of all data together
write.csv(allRes_NNETAR_output_all, "../Data/Forecasts/NNETAR_4casts/allReservoir_NNETAR_4cast.csv", row.names = F)


```



## Compile Climatology 4cast outputs  

```{r}
#FCR Clim forecasts
#bind forecast outputs into one data frame 
fcr_Clim_output_all <- list.files(path = "../Data/Forecasts/Clim_4casts/fcre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#BVR Clim forecasts
#bind forecast outputs into one data frame 
bvr_Clim_output_all <- list.files(path = "../Data/Forecasts/Clim_4casts/bvre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#CCR Clim forecasts
#bind forecast outputs into one data frame 
ccr_Clim_output_all <- list.files(path = "../Data/Forecasts/Clim_4casts/ccre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#Compile Clim forecasts and write csv
allres_Clim <- rbind(fcr_Clim_output_all, bvr_Clim_output_all, ccr_Clim_output_all)

write.csv(allres_Clim, "../Data/Forecasts/Clim_4casts/allReservoir_Clim_4cast.csv", row.names = F)


```



## Compile Persistence 4cast outputs  

```{r}
#FCR Persist forecasts
#bind forecast outputs into one data frame 
fcr_Persist_output_all <- list.files(path = "../Data/Forecasts/Persist_4casts/fcre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#BVR Persist forecasts
#bind forecast outputs into one data frame 
bvr_Persist_output_all <- list.files(path = "../Data/Forecasts/Persist_4casts/bvre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#CCR Persist forecasts
#bind forecast outputs into one data frame 
ccr_Persist_output_all <- list.files(path = "../Data/Forecasts/Persist_4casts/ccre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#Compile Persist forecasts and write csv
allres_Persist <- rbind(fcr_Persist_output_all, bvr_Persist_output_all, ccr_Persist_output_all)

write.csv(allres_Persist, "../Data/Forecasts/Persist_4casts/allReservoir_Persist_4cast.csv", row.names = F)

```



## Compile RainLag 4cast outputs  

```{r}
#FCR RainLab forecasts
#bind forecast outputs into one data frame 
fcr_RainLag_output_all <- list.files(path = "../Data/Forecasts/RainLag_4casts/fcre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#BVR RainLab forecasts
#bind forecast outputs into one data frame 
bvr_RainLag_output_all <- list.files(path = "../Data/Forecasts/RainLag_4casts/bvre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#CCR RainLab forecasts
#bind forecast outputs into one data frame 
ccr_RainLag_output_all <- list.files(path = "../Data/Forecasts/RainLag_4casts/ccre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  

#Compile RainLag forecasts and write csv
allres_RainLag <- rbind(fcr_RainLag_output_all, bvr_RainLag_output_all, ccr_RainLag_output_all)

write.csv(allres_RainLag, "../Data/Forecasts/RainLag_4casts/allReservoir_RainLag_4cast.csv", row.names = F)

```



## Compile Meteo Regression 4cast outputs  

```{r}
#FCR MetRegress forecasts
#bind forecast outputs into one data frame 
fcr_MetRegress_output_all <- list.files(path = "../Data/Forecasts/MetRegress_4casts/fcre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()  


#BVR MetRegress forecasts
#bind forecast outputs into one data frame 
bvr_MetRegress_output_all <- list.files(path = "../Data/Forecasts/MetRegress_4casts/bvre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()


#CCR MetRegress forecasts
#bind forecast outputs into one data frame 
ccr_MetRegress_output_all <- list.files(path = "../Data/Forecasts/MetRegress_4casts/ccre", pattern = "*.csv", full.names = T) |> 
  base::lapply(read_csv) |> 
  bind_rows()


#Compile MetRegress forecasts and write csv
allres_MetRegress <- rbind(fcr_MetRegress_output_all, bvr_MetRegress_output_all, ccr_MetRegress_output_all)

write.csv(allres_MetRegress, "../Data/Forecasts/MetRegress_4casts/allReservoir_MetRegress_4cast.csv", row.names = F)


```


## Scoring forecast 

read in all forecasts csvs
```{r}
#AR
allres_AR <- read.csv("../Data/Forecasts/AR_4casts/allReservoir_AR_4cast.csv")

#NNAR
allRes_NNETAR_output_all <- read.csv("../Data/Forecasts/NNETAR_4casts/allReservoir_NNETAR_4cast.csv")

#Met Regress
allres_MetRegress <- read.csv("../Data/Forecasts/MetRegress_4casts/allReservoir_MetRegress_4cast.csv")

#Rain lag
allres_RainLag <- read.csv("../Data/Forecasts/RainLag_4casts/allReservoir_RainLag_4cast.csv")

#Clim
allres_Clim <- read.csv("../Data/Forecasts/Clim_4casts/allReservoir_Clim_4cast.csv")

#Persist
allres_Persist <- read.csv("../Data/Forecasts/Persist_4casts/allReservoir_Persist_4cast.csv")


```


using score4cast to get crps
```{r}
##uncomment if folders don't exist locally
# dir.create("../Data/Scores")

## AR scores 
AR_scores <- crps_logs_score(allres_AR, allres_targets)

write.csv(AR_scores, "../Data/Scores/allReservoir_AR_4cast_Scores.csv", row.names = F)

## RainLag scores 
RainLag_scores <- crps_logs_score(allres_RainLag, allres_targets)

write.csv(RainLag_scores, "../Data/Scores/allReservoir_RainLag_4cast_Scores.csv", row.names = F)

## MetRegress scores 
MetRegress_scores <- crps_logs_score(allres_MetRegress, allres_targets)

write.csv(MetRegress_scores, "../Data/Scores/allReservoir_MetRegress_4cast_Scores.csv", row.names = F)

## nnetar scores 
nnetar_scores <- crps_logs_score(allRes_NNETAR_output_all, allres_targets)

write.csv(nnetar_scores, "../Data/Scores/allReservoir_NNETAR_4cast_Scores.csv", row.names = F)

## Clim scores 
Clim_scores <- crps_logs_score(allres_Clim, allres_targets)

write.csv(Clim_scores, "../Data/Scores/allReservoir_Clim_4cast_Scores.csv", row.names = F)

## Persist scores 
Persist_scores <- crps_logs_score(allres_Persist, allres_targets)

write.csv(Persist_scores, "../Data/Scores/allReservoir_Persist_4cast_Scores.csv", row.names = F)


```

Bind all scores together and write csv

```{r}
##Bind scores together 
All_Scores <- rbind(AR_scores, RainLag_scores, MetRegress_scores, nnetar_scores, Clim_scores, Persist_scores)  

write.csv(All_Scores, "../Data/Scores/allReservoir_allmodel_4cast_Scores_4apr25.csv", row.names = F)

```
