==> Scripts/4_Analyze_4casts.Rmd <==
---
title: "Analyze VA reservoir fDOM 4casts"
author: "Dexter Howard"
date: "2024-09-06"
output: html_document
---

```{r}
library(tidyverse)
library(patchwork)
```



## Read in data and format 

```{r}
#read in scores 
Scored_forecasts <- read_csv("../Data/allReservoir_allmodel_4cast_Scores_4apr25.csv") |> 
  mutate(Horizon = datetime - reference_datetime) |> 
  filter(Horizon > 0, Horizon < 35) |>
  mutate(site_id = ifelse(site_id == "bvre", "BVR", site_id),
         site_id = ifelse(site_id == "fcre", "FCR", site_id),
         site_id = ifelse(site_id == "ccre", "CCR", site_id)) |> 
  mutate(model_id = ifelse(model_id == "fable_persistence", "Persistence", model_id),
         model_id = ifelse(model_id == "climatology", "DOY", model_id),
         model_id = ifelse(model_id == "fDOM_AR_dwh", "TS_AR", model_id),
         model_id = ifelse(model_id == "fDOM_RainLag_dwh", "TS_Rain", model_id),
         model_id = ifelse(model_id == "fDOM_MetRegression_dwh", "TS_Met", model_id),
         model_id = ifelse(model_id == "fableNNETAR", "NNAR", model_id)
         ) |>  #update model names for plotting
    mutate(model_type = ifelse(model_id %in% c("Persistence", "DOY"), "Baseline", "Data Driven")
         ) #grouping by model types


#set model colors and shapes
model_colors <- c("TS_AR" = "#CC6677", "DOY" = "#88CCEE", "Persistence" = "#DDCC77", "NNAR" = "#117733", "TS_Rain" = "#332288", "TS_Met" = "#888888")
model_shape <- c("Data Driven" = 1, "Baseline" = 4)

```


## Checks on missing data and forecasts 


number of missing AR day 

```{r}

AR <- Scored_forecasts |> filter(model_id == "TS_AR")

#no observation
AR_na <- AR |> filter(is.na(observation))
missdateslist <- unique(AR_na$datetime)
AR_na2 <- AR_na |> filter(datetime %in% missdateslist)


```



## RMSE and Skills Plots 


RMSE and faceted metrics over horizon

```{r}

## Set up data for RMSE horizon plot
a <- Scored_forecasts |> 
  group_by(Horizon, site_id, model_id, model_type) |> 
  summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 1)) 

# add a panel for across reservoirs 
b <- Scored_forecasts |> 
  group_by(Horizon, model_id, model_type) |> 
  summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 1)) |> 
  mutate(site_id = "Across Sites")


## Set up data for all metrics SI figure
c <- Scored_forecasts |> 
   group_by(site_id, model_id, model_type, Horizon) |> 
   summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2),
             MAE = mean(abs(mean - observation), na.rm = T),
             CRPS = mean(crps, na.rm = T),
             SD = mean(sd, na.rm = T)
            ) |> 
  pivot_longer(-c(1:4))

# make panel for all sites average
d <- Scored_forecasts |> 
   group_by(model_id, model_type, Horizon) |> 
   summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2),
             MAE = mean(abs(mean - observation), na.rm = T),
             CRPS = mean(crps, na.rm = T),
             SD = mean(sd, na.rm = T)
            ) |> 
  pivot_longer(-c(1:3)) |> 
  mutate(site_id = "Across Sites")


#### Make SI figure 

SI_all_metrics_over_horizon <- rbind(c,d) |> 
  mutate(site_id = ifelse(site_id == "Across Sites", "Across Reservoirs", site_id)) |> 
  filter(!name %in% c("MAE")) |> 
ggplot(aes(x = Horizon, y = value, color = model_id, linetype = model_type))+
  geom_line(linewidth = 1.2)+
  facet_grid(name~factor(site_id, levels = c("Across Reservoirs","BVR", "CCR", "FCR")), scales = "free_y")+
  labs(x = "Horizon", y = "Metric (QSU)")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 12))+
    # scale_x_continuous(breaks = c(1,4,7,10,13, 16))+
  guides(linetype = guide_legend(nrow = 2))+
  theme(legend.text = element_text(size = 8), legend.title = element_text(size = 10),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = model_colors)+  scale_linetype_manual(values = model_shape)

SI_all_metrics_over_horizon

#ggsave("../Figures/figS2_allmetrics_Horizon.png", SI_all_metrics_over_horizon, height = 6, width = 6, units = "in")

```



Skill and associated SI figure 

```{r}

#### Set up data frames by metric and null 

##CRPS Persist
#by reservoir
skill_crps_persist_a <- Scored_forecasts |> 
  select(reference_datetime, datetime, Horizon, site_id, model_id, crps) |> 
  pivot_wider(names_from = model_id, values_from = crps) |> 
  mutate(Skill_AR_crps_P = 1-(TS_AR/Persistence),
         Skill_NNETAR_crps_P = 1-(NNAR/Persistence),
         Skill_MetRegress_crps_P = 1-(TS_Met/Persistence),
         Skill_RainLag_crps_P = 1-(TS_Rain/Persistence)) |> 
  select(site_id, Horizon, Skill_AR_crps_P:Skill_RainLag_crps_P) |> 
  pivot_longer(-c(1:2)) |> 
  group_by(site_id, Horizon, name) |> 
  summarise(value = mean(value, na.rm = T))

#across sites
skill_crps_persist_b <- Scored_forecasts |> 
  select(reference_datetime, datetime, Horizon, site_id, model_id, crps) |> 
  pivot_wider(names_from = model_id, values_from = crps) |> 
  mutate(Skill_AR_crps_P = 1-(TS_AR/Persistence),
         Skill_NNETAR_crps_P = 1-(NNAR/Persistence),
         Skill_MetRegress_crps_P = 1-(TS_Met/Persistence),
         Skill_RainLag_crps_P = 1-(TS_Rain/Persistence)) |> 
  select(site_id, Horizon, Skill_AR_crps_P:Skill_RainLag_crps_P) |> 
  pivot_longer(-c(1:2)) |> 
  group_by(Horizon, name) |> 
  summarise(value = mean(value, na.rm = T)) |> 
  mutate(site_id = "Across Reservoirs")

#bind
skill_crps_persist <- rbind(skill_crps_persist_a, skill_crps_persist_b)

##CRPS Clim
#by res
skill_crps_clim_a <- Scored_forecasts |> 
  select(reference_datetime, datetime, Horizon, site_id, model_id, crps) |> 
  pivot_wider(names_from = model_id, values_from = crps) |> 
  mutate(Skill_AR_crps_C = 1-(TS_AR/DOY),
         Skill_NNETAR_crps_C = 1-(NNAR/DOY),
         Skill_MetRegress_crps_C = 1-(TS_Met/DOY),
         Skill_RainLag_crps_C = 1-(TS_Rain/DOY)) |> 
  select(site_id, Horizon, Skill_AR_crps_C:Skill_RainLag_crps_C) |> 
  pivot_longer(-c(1:2)) |> 
  group_by(site_id, Horizon, name) |> 
  summarise(value = mean(value, na.rm = T))

#across sites
skill_crps_clim_b <- Scored_forecasts |> 
  select(reference_datetime, datetime, Horizon, site_id, model_id, crps) |> 
  pivot_wider(names_from = model_id, values_from = crps) |> 
  mutate(Skill_AR_crps_C = 1-(TS_AR/DOY),
         Skill_NNETAR_crps_C = 1-(NNAR/DOY),
         Skill_MetRegress_crps_C = 1-(TS_Met/DOY),
         Skill_RainLag_crps_C = 1-(TS_Rain/DOY)) |> 
  select(site_id, Horizon, Skill_AR_crps_C:Skill_RainLag_crps_C) |> 
  pivot_longer(-c(1:2)) |> 
  group_by(Horizon, name) |> 
  summarise(value = mean(value, na.rm = T)) |> 
  mutate(site_id = "Across Reservoirs")

skill_crps_clim <- rbind(skill_crps_clim_a, skill_crps_clim_b)

##RMSE Persist
#by res
skill_rmse_persist_a <- Scored_forecasts |> 
   group_by(site_id, model_id, Horizon) |> 
   summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2)
            ) |> 
  select(Horizon, site_id, model_id, RMSE) |> 
  pivot_wider(names_from = model_id, values_from = RMSE) |> 
  mutate(Skill_AR_rmse_P = 1-(TS_AR/Persistence),
         Skill_NNETAR_rmse_P = 1-(NNAR/Persistence),
         Skill_MetRegress_rmse_P = 1-(TS_Met/Persistence),
         Skill_RainLag_rmse_P = 1-(TS_Rain/Persistence)) |> 
  select(site_id, Horizon, Skill_AR_rmse_P:Skill_RainLag_rmse_P) |> 
  pivot_longer(-c(1:2))

#acorss sites
skill_rmse_persist_b <- Scored_forecasts |> 
   group_by(model_id, Horizon) |> 
   summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2)
            ) |> 
  select(Horizon, model_id, RMSE) |> 
  pivot_wider(names_from = model_id, values_from = RMSE) |> 
  mutate(Skill_AR_rmse_P = 1-(TS_AR/Persistence),
         Skill_NNETAR_rmse_P = 1-(NNAR/Persistence),
         Skill_MetRegress_rmse_P = 1-(TS_Met/Persistence),
         Skill_RainLag_rmse_P = 1-(TS_Rain/Persistence)) |> 
  select(Horizon, Skill_AR_rmse_P:Skill_RainLag_rmse_P) |> 
  pivot_longer(-c(1)) |> 
  mutate(site_id = "Across Reservoirs")

skill_rmse_persist <- rbind(skill_rmse_persist_a, skill_rmse_persist_b)

##RMSE Clim
#by res
skill_rmse_clim_a <- Scored_forecasts |> 
   group_by(site_id, model_id, Horizon) |> 
   summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2)
            ) |> 
  select(Horizon, site_id, model_id, RMSE) |> 
  pivot_wider(names_from = model_id, values_from = RMSE) |> 
  mutate(Skill_AR_rmse_C = 1-(TS_AR/DOY),
         Skill_NNETAR_rmse_C = 1-(NNAR/DOY),
         Skill_MetRegress_rmse_C = 1-(TS_Met/DOY),
         Skill_RainLag_rmse_C = 1-(TS_Rain/DOY)) |> 
  select(site_id, Horizon, Skill_AR_rmse_C:Skill_RainLag_rmse_C) |> 
  pivot_longer(-c(1:2))

#across sites
skill_rmse_clim_b <- Scored_forecasts |> 
   group_by(model_id, Horizon) |> 
   summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2)
            ) |> 
  select(Horizon, model_id, RMSE) |> 
  pivot_wider(names_from = model_id, values_from = RMSE) |> 
  mutate(Skill_AR_rmse_C = 1-(TS_AR/DOY),
         Skill_NNETAR_rmse_C = 1-(NNAR/DOY),
         Skill_MetRegress_rmse_C = 1-(TS_Met/DOY),
         Skill_RainLag_rmse_C = 1-(TS_Rain/DOY)) |> 
  select(Horizon, Skill_AR_rmse_C:Skill_RainLag_rmse_C) |> 
  pivot_longer(-c(1))|> 
  mutate(site_id = "Across Reservoirs")

skill_rmse_clim <- rbind(skill_rmse_clim_a, skill_rmse_clim_b)


#### Put data frames together 
skill_df <- rbind(skill_crps_clim, skill_crps_persist, skill_rmse_clim, skill_rmse_persist) |> 
  mutate(name_a = name) |>  #make a copy of name for reference
  select(site_id, Horizon, name_a, name, value) |> 
  tidyr::separate(name, into = c("Skill", "model_id", "metric", "Null"), sep = "_") |>  #break apart name column for grouping 
  mutate(Null = ifelse(Null == "P", "Skill_Peristence", "Skill_Climatology")) |> 
  mutate(model_idA = ifelse(model_id == "AR", "TS_AR", NA),
         model_idA = ifelse(model_id == "NNETAR", "NNAR", model_idA),
         model_idA = ifelse(model_id == "MetRegress", "TS_Met", model_idA),
         model_idA = ifelse(model_id == "RainLag", "TS_Rain", model_idA)) #fix model names to match above




#### Make plots 
#Skill CRPS 
skill_crps_plot <- skill_df |> 
  filter(metric == "crps") |> 
  mutate(NullA = ifelse(Null == "Skill_Climatology", "DOY", "Persistence")) |> 
  #filter(!site_id == "Across Sites") |> 
  ggplot(aes(x = Horizon, y = value, color = model_idA))+
  geom_line()+
  facet_grid(NullA~factor(site_id, levels = c("Across Reservoirs", "BVR", "CCR", "FCR")), scales = "free_y")+
  geom_hline(yintercept = 0, linewidth =1.2, linetype = 2)+
  geom_line(linewidth = 1.2)+
  # scale_x_continuous(breaks = c(1,4,7,10,13,16))+ 
  guides(colour = guide_legend(nrow = 1))+
  labs(x = "Horizon", color = "model", title = "b) CRPS Skill")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = model_colors)

#Skill RMSE
skill_rmse_plot <- skill_df |> 
  filter(metric == "rmse") |> 
  mutate(NullA = ifelse(Null == "Skill_Climatology", "DOY", "Persistence")) |> 
  #filter(!site_id == "Across Sites") |> 
  ggplot(aes(x = Horizon, y = value, color = model_idA))+
  geom_line()+
  facet_grid(NullA~factor(site_id, levels = c("Across Reservoirs", "BVR", "CCR", "FCR")), scales = "free_y")+
  geom_hline(yintercept = 0, linewidth =1.2, linetype = 2)+
  geom_line(linewidth = 1.2)+
  # scale_x_continuous(breaks = c(1,4,7,10,13,16))+ 
  guides(colour = guide_legend(nrow = 1))+
  labs(x = "Horizon", color = "model", title = "a) RMSE Skill")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = model_colors)


#panel 

all_skill_SI <- skill_rmse_plot / skill_crps_plot

all_skill_SI
#ggsave("../Figures/figS3_all_skill_SI.png", all_skill_SI, height = 8, width = 6.5, units = "in")


```

RMSE combined plot 

```{r}

#RMSE horizon plot
rmse_for_join <- rbind(a,b) |> 
    mutate(site_id = ifelse(site_id == "Across Sites", "Across Reservoirs", site_id)) |> 
  ggplot(aes(x = Horizon, y = RMSE, color = model_id, linetype = model_type))+ 
  geom_line(linewidth = 1.2)+
  facet_wrap(~factor(site_id, levels = c("Across Reservoirs", "BVR", "CCR", "FCR")), scales = "fixed", ncol = 4)+

  labs(x = "Horizon", y = "RMSE (QSU)", color = "Model ID", linetype = "Model Type")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 12),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  guides(linetype = guide_legend(nrow = 2))+
  scale_color_manual(values = model_colors)+  scale_linetype_manual(values = model_shape)

#RMSE persistence skill plot
skill_for_join <- skill_df |> 
  filter(metric == "rmse",
         Null == "Skill_Peristence") |> 
  #filter(!site_id == "Across Sites") |> 
  ggplot(aes(x = Horizon, y = value, color = model_idA))+
  geom_line()+
  facet_grid(~factor(site_id, levels = c("Across Reservoirs", "BVR", "CCR", "FCR")), scales = "free_y")+
  geom_hline(yintercept = 0, linewidth =1.2, linetype = 2)+
  geom_line(linewidth = 1.2)+
  labs(x = "Horizon", y = "Forecast Skill", color = "Model ID")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 12),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = model_colors)

rmse_skill_combo <- ggpubr::ggarrange(rmse_for_join, skill_for_join, ncol = 1, labels = "auto")

rmse_skill_combo

#ggsave("../Figures/fig5_RMSE_and_Skill_combo_MS_longer.png", rmse_skill_combo, height = 8, width = 6.1, units = "in")



```

EFI plots
```{r}
efirmse <- rbind(a,b) |> 
    mutate(site_id = ifelse(site_id == "Across Sites", "Across Reservoirs", site_id)) |> 
  ggplot(aes(x = Horizon, y = RMSE, color = model_id, linetype = model_type))+ 
  geom_line(linewidth = 2)+
  facet_wrap(~factor(site_id, levels = c("Across Reservoirs", "BVR", "CCR", "FCR")), scales = "fixed", ncol = 1)+

  labs(x = "Horizon", y = "RMSE (QSU)", color = "Model ID", linetype = "Model Type")+
  theme_bw() + theme(legend.position = "none", text = element_text(size = 48),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  guides(linetype = guide_legend(nrow = 2))+
  scale_color_manual(values = model_colors)+  scale_linetype_manual(values = model_shape)


ggsave("../Figures/EFI_rmse_long.png", efirmse, height = 18, width = 14, units = "in")


efiskill <- skill_df |> 
  filter(metric == "rmse",
         Null == "Skill_Peristence") |> 
  #filter(!site_id == "Across Sites") |> 
  ggplot(aes(x = Horizon, y = value, color = model_idA))+
  geom_line()+
  facet_grid(~factor(site_id, levels = c("Across Reservoirs", "BVR", "CCR", "FCR")), scales = "free_y")+
  geom_hline(yintercept = 0, linewidth =2, linetype = 2)+
  geom_line(linewidth = 2)+
  labs(x = "Horizon", y = "Forecast Skill", color = "Model ID")+
  theme_bw() + theme(legend.position = "none", text = element_text(size = 36),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = model_colors)

ggsave("../Figures/EFI_skill.png", efiskill, height = 8, width = 17, units = "in")



```



## RMSE and Skill Stats


```{r}

#look at interactive plots

plotly::ggplotly(skill_rmse_plot)
plotly::ggplotly(skill_crps_plot)

plotly::ggplotly(rmse_for_join)
plotly::ggplotly(skill_for_join)


```




Make tables for MS: Table 2 (RMSE) and S1 (CRPS); and SD stats for Supp Text S1

```{r}

########################################## RMSE  #

#table to get metrics by horizon 
rbind(c,d) |> 
  filter(name %in% c("RMSE", "CRPS"),
         Horizon %in% c(1,7,14,28)) |> 
  pivot_wider(names_from = name, values_from = value)

rbind(c,d) |> 
  filter(name %in% c("RMSE"),
         Horizon %in% c(1,7,14,28)) |> 
  pivot_wider(names_from = Horizon, values_from = value)

###Tables from first submission 

#across sites and all models 
rbind(c,d) |> 
  group_by(site_id, model_id, name) |> 
  filter(name == "RMSE") |> 
  summarise(min = round(min(value, na.rm = T),1),
            mean = round(mean(value, na.rm = T),1), median = round(median(value, na.rm = T),1),
            max = round(max(value, na.rm = T),1),        range = max - min    ) |> 
  select(site_id, model_id, name, median, min, max, range)


#across sites and models 
rbind(c,d) |> 
  group_by(site_id, name) |> 
  filter(name == "RMSE") |> 
  summarise(min = round(min(value, na.rm = T),1),
            mean = round(mean(value, na.rm = T),1), median = round(median(value, na.rm = T),1),
            max = round(max(value, na.rm = T),1),        range = max - min    ) |> 
  select(site_id, name, median, min, max, range)


############################## CRPS #
#across sites and all models 
rbind(c,d) |> 
  group_by(site_id, model_id, name) |> 
  filter(name == "CRPS") |> 
  summarise(min = round(min(value, na.rm = T),1),
            mean = round(mean(value, na.rm = T),1), median = round(median(value, na.rm = T),1),
            max = round(max(value, na.rm = T),1),        range = max - min    ) |> 
  select(site_id, model_id, name, median, min, max, range)


#across sites and models 
rbind(c,d) |> 
  group_by(site_id, name) |> 
  filter(name == "CRPS") |> 
  summarise(min = round(min(value, na.rm = T),1),
            mean = round(mean(value, na.rm = T),1), median = round(median(value, na.rm = T),1),
            max = round(max(value, na.rm = T),1),        range = max - min    ) |> 
  select(site_id, name, median, min, max, range)


############################## SD #
#across sites and all models 
rbind(c,d) |> 
  group_by(site_id, model_id, name) |> 
  filter(name == "SD") |> 
  summarise(min = round(min(value, na.rm = T),1),
            mean = round(mean(value, na.rm = T),1), median = round(median(value, na.rm = T),1),
            max = round(max(value, na.rm = T),1),        range = max - min    ) |> 
  select(site_id, model_id, name, median, min, max, range)


#across sites and models 
rbind(c,d) |> 
  group_by(site_id, name) |> 
  filter(name == "SD") |> 
  summarise(min = round(min(value, na.rm = T),1),
            mean = round(mean(value, na.rm = T),1), median = round(median(value, na.rm = T),1),
            max = round(max(value, na.rm = T),1),        range = max - min    ) |> 
  select(site_id, name, median, min, max, range)


```




## Seasons Figure and Stats

```{r}
#### make seasons groupings 
Scored_forecasts_seasons <- Scored_forecasts |> 
   mutate(month = month(datetime),
         Season2 = ifelse(month %in% c(3,4,5), "Spring", NA),
         Season2 = ifelse(month %in% c(6,7,8), "Summer", Season2),
         Season2 = ifelse(month %in% c(9,10,11), "Fall", Season2),
         Season2 = ifelse(month %in% c(1,2,12), "Winter", Season2)
         ) 



####Get data for seasons with across reservoirs panel 

e <- Scored_forecasts_seasons |> 
  group_by(Horizon, site_id, model_id, model_type, Season2) |> 
  summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2))

f <- Scored_forecasts_seasons |> 
  group_by(Horizon, model_id, model_type, Season2) |> 
  summarise(RMSE = round(sqrt(mean((mean - observation)^2, na.rm = TRUE)), 2)) |> 
  mutate(site_id = "Across Reservoirs")




## SI figure all models across seasons and sites

rmse_season_v2 <- rbind(e,f) |> 
  ggplot(aes(x = Horizon, y = RMSE, color = model_id, linetype = model_type))+ # color = model_id
  geom_line(linewidth = 1.2)+
  facet_grid(factor(site_id, levels = c("BVR", "CCR", "FCR", "Across Reservoirs"))
             ~
               factor(Season2, levels = c("Winter", "Spring", "Summer", "Fall" )))+
  # scale_x_continuous(breaks = c(1,4,7,10,13,16))+
  labs(x = "Horizon", y = "RMSE (QSU)")+
  guides(linetype = guide_legend(nrow = 2))+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = model_colors)+ scale_linetype_manual(values = model_shape)

rmse_season_v2

#ggsave("../Figures/Fig_S4_rmse_season_SI.png", rmse_season_v2, height = 7, width = 6.5, units = "in")



#### MS figure across models by season and site
rmse_seasons_avgModels <- rbind(e,f) |> 
  ungroup() |> 
  group_by(Horizon, site_id, Season2) |> 
  reframe(RMSE = mean(RMSE, na.rm = T)) |> 
  ggplot(aes(x = Horizon, y = RMSE, color = site_id))+ 
  geom_line(linewidth = 1.2)+
  facet_wrap(~factor(Season2, levels = c("Winter", "Spring", "Summer", "Fall" )), ncol = 4)+
  # scale_x_continuous(breaks = c(1,4,7,10,13,16))+
  labs(x = "Horizon", y = "RMSE (QSU)", color = "Reservoir")+
  guides(linetype = guide_legend(nrow = 2))+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = c("FCR" = "#009E73", "BVR" = "orange", "CCR" = "skyblue", "Across Reservoirs" = "black"))

rmse_seasons_avgModels

#ggsave("../Figures/Fig6_rmse_season_modelAVG.png", rmse_seasons_avgModels, height = 3.5, width = 6, units = "in")



```


Seasons stats 

```{r}

##### stats by reservoir, across reservoir for seasons  
rbind(e,f) |> 
  ungroup() |> 
  group_by(Horizon, site_id, Season2) |> 
  reframe(RMSE = mean(RMSE, na.rm = T)) |> 
  group_by(site_id, Season2) |> 
  summarise(min = round(min(RMSE, na.rm = T),1),
            mean = round(mean(RMSE, na.rm = T),1),      median = round(median(RMSE, na.rm = T),1),
            max = round(max(RMSE, na.rm = T),1),      range = max - min        )


# look at range across models by horizon
plotly::ggplotly(rmse_season_v2)

```

## 95% CI interval stats

```{r}
head(Scored_forecasts)

CI95_stats <- Scored_forecasts |> 
  mutate(CI95 = ifelse(observation <= quantile97.5 & observation >= quantile02.5, 1, 0 )) |> 
  group_by(model_id, model_type, site_id, Horizon) |> 
  summarise(CI95_sum = sum(CI95, na.rm = T),
            CI95_n = sum(!is.na(CI95))) |> 
  mutate(CI95_percent = CI95_sum/CI95_n)


CI95fig <- CI95_stats |> 
  ggplot(aes(x = Horizon, y = (CI95_percent)*100, color = model_id, linetype = model_type))+
  geom_line(size = 1.5)+
  facet_wrap(~site_id)+
  geom_hline(yintercept = 95, linetype = 1, size = 0.75)+
  labs(x = "Forecast Horizon (days)", y = "Percent of observations \n within the 95% CI")+
  scale_color_manual(values = model_colors)+
scale_linetype_manual(values = model_shape)+
  guides(linetype = guide_legend(nrow = 2))+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                     panel.grid.major = element_blank(), panel.grid.minor = element_blank())

CI95fig

ggsave("../Figures/SI_95percentCI.png", CI95fig, height = 4, width = 7.5, units = "in")

  


```

## full TS of forecasts

```{r}
## 1 day
fullts_1day <- Scored_forecasts |> 
  filter(Horizon %in% c(1)) |> 
  # mutate(Horizon_label = ifelse(Horizon == 1, "1 day ahead", NA)) |> 
  ggplot()+
    geom_point(aes(x = as.Date(datetime), y = observation), color = "red", size = 0.5)+
  geom_ribbon(aes(x = as.Date(datetime), ymin = quantile02.5, ymax = quantile97.5), fill = "gray", alpha = 0.5)+
  geom_line(aes(x = as.Date(datetime), y = mean))+
  facet_grid(site_id~model_id, scales = "free_y")+
  #ylim(5,30)+
  scale_x_date(date_breaks = "4 month", date_labels = "%b" )+
  labs(x = element_blank(), y = "fDOM (QSU)", title = "1 day ahead forecasts")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                     panel.grid.major = element_blank(), panel.grid.minor = element_blank())

## 16 day
fullts_16day <- Scored_forecasts |> 
  filter(Horizon %in% c(16)) |> 
  ggplot()+
  geom_point(aes(x = as.Date(datetime), y = observation), color = "red", size = 0.5)+
  geom_ribbon(aes(x = as.Date(datetime), ymin = quantile02.5, ymax = quantile97.5), fill = "gray", alpha = 0.5)+
  geom_line(aes(x = as.Date(datetime), y = mean))+
  facet_grid(site_id~model_id, scales = "free_y")+
  #ylim(5,30)+
  scale_x_date(date_breaks = "4 month", date_labels = "%b" )+
  labs(x = element_blank(), y = "fDOM (QSU)", title = "16 day ahead forecasts")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                     panel.grid.major = element_blank(), panel.grid.minor = element_blank())


## 28 day
fullts_34day <- Scored_forecasts |> 
  filter(Horizon %in% c(34)) |> 
  ggplot()+
  geom_point(aes(x = as.Date(datetime), y = observation), color = "red", size = 0.5)+
  geom_ribbon(aes(x = as.Date(datetime), ymin = quantile02.5, ymax = quantile97.5), fill = "gray", alpha = 0.5)+
  geom_line(aes(x = as.Date(datetime), y = mean))+
  facet_grid(site_id~model_id, scales = "free_y")+
  #ylim(5,30)+
  scale_x_date(date_breaks = "4 month", date_labels = "%b" )+
  labs(x = element_blank(), y = "fDOM (QSU)", title = "34 day ahead forecasts")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                     panel.grid.major = element_blank(), panel.grid.minor = element_blank())


SIfig_full_4castTS <- fullts_1day / fullts_16day / fullts_34day

#ggsave("../Figures/SIfig_full_4castTS_v2.png", SIfig_full_4castTS, height = 8, width = 6.5, units = "in")


```



## Observed fDOM TS

Plots of observed daily data 

```{r}
#all on one plot
targets <- read_csv("../Data/GeneratedData/Targets_fDOM_allReservoirs.csv")

fdom_TS <- targets |> 
  filter(datetime >= ymd("2023-03-13"),
         datetime <= ymd("2025-04-03")
         ) |> 
  mutate(site_id_A = ifelse(site_id == "fcre", "FCR", NA),
         site_id_A = ifelse(site_id == "bvre", "BVR", site_id_A),
         site_id_A = ifelse(site_id == "ccre", "CCR", site_id_A)) |> 
  ggplot(aes(x = datetime, y = observation, color = site_id_A))+
  geom_vline(xintercept = ymd("2024-03-01"), linetype = 2, linewidth = 1.2)+
    annotate("text", x = as.Date("2023-08-15"), y = 28, label = "Calibration", 
             color = "black", size = 4, fontface = "bold")+
    annotate("text", x = as.Date("2024-05-15"), y = 28, label = "Forecasting", 
             color = "black", size = 4, fontface = "bold")+
  geom_point()+
  scale_y_continuous(limits = c(3,28), breaks = c(5,10,15,20,25))+
  scale_x_date(date_labels = "%b", breaks = "2 months")+
  labs(x = "Date", y = "fDOM (QSU)", color = "Reservoir")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 18),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = c("FCR" = "#009E73", "BVR" = "orange", "CCR" = "skyblue"))

fdom_TS

#ggsave("../Figures/Fig4_Observed_FDOM_TS.png", fdom_TS, height = 4, width = 6.5, units = "in")



#### Turnover and Helene plot (calling Helene landfall 024-09-26)

turnover_helene_plot <- targets |>
  filter(datetime >= ymd("2024-09-01"),
         datetime < ymd("2024-12-01") ) |>
  mutate(site_id_A = ifelse(site_id == "fcre", "FCR", NA),
         site_id_A = ifelse(site_id == "bvre", "BVR", site_id_A),
         site_id_A = ifelse(site_id == "ccre", "CCR", site_id_A)) |>
  ggplot(aes(x = datetime, y = observation, color = site_id_A))+
  geom_point()+ geom_vline(xintercept = ymd("2024-09-26"))+
  annotate(geom = "point", x = ymd("2024-10-28"), y = 22, shape = 15, size = 3, color = "#009E73")+
  annotate(geom = "point", x = ymd("2024-11-16"), y = 22, shape = 15, size = 3, color = "orange")+
  annotate(geom = "point", x = ymd("2024-11-03"), y = 22, shape = 15, size = 3, color = "skyblue")+
  labs(x = "Date", y = "fDOM (QSU)", color = "Reservoir")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 18),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = c("FCR" = "#009E73", "BVR" = "orange", "CCR" = "skyblue"))

turnover_helene_plot
plotly::ggplotly(turnover_helene_plot)

###Helene numbers
##CCR fDOM level rises from 6.7 to 12.8 QSU (6.7 QSU) between 25sep24 (day before) and 6oct24 (10 days after)
##BVR fDOM level rises from 14.2 to 15.4 QSU (1.2 QSU) between 25sep24 (day before) and 6oct24 (10 days after)
##FCR fDOM level rises from 17.6 to 26.3 QSU (8.7 QSU) between 25sep24 (day before) and 6oct24 (10 days after)

###Turnover numbers


##### FEb 2025 storm plots 
febstorm_plot <- targets |>
  filter(datetime >= ymd("2025-01-15"),
         datetime < ymd("2025-02-28") ) |>
  mutate(site_id_A = ifelse(site_id == "fcre", "FCR", NA),
         site_id_A = ifelse(site_id == "bvre", "BVR", site_id_A),
         site_id_A = ifelse(site_id == "ccre", "CCR", site_id_A)) |>
  ggplot(aes(x = datetime, y = observation, color = site_id_A))+
  geom_point()+ geom_vline(xintercept = ymd("2025-02-10"))+
  labs(x = "Date", y = "fDOM (QSU)", color = "Reservoir")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 18),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = c("FCR" = "#009E73", "BVR" = "orange", "CCR" = "skyblue"))

febstorm_plot
plotly::ggplotly(febstorm_plot)

###Feb storm numbers
##CCR fDOM level rises from 12.9 to 16.7 QSU (3.8 QSU), w/ max of 18.6 (5.7) between 11feb25 (day before) and 21feb25 (10 days after)
##BVR fDOM level rises from 10.6 to 10.4 QSU w/ min of 10.1 between 11feb25 (day before) and 21feb25 (10 days after)
##FCR fDOM level rises from 14.1 to 17.0 QSU (2.9 QSU) w/ max of 18.2 between 11feb25 (day before) and 21feb25 (10 days after)



```

Summary stats of observed data

```{r}
#calibration period 
targets |> 
  filter(datetime >= ymd("2023-03-13"),
         datetime <= ymd("2024-02-28")
         ) |> 
  group_by(site_id) |> 
  summarise(min = round(min(observation, na.rm = T), 1),
            median = round(median(observation, na.rm = T), 1),
            mean = round(mean(observation, na.rm = T), 1),
            max = round(max(observation, na.rm = T), 1),
            sd = round(sd(observation, na.rm = T), 1),
            range = max - min) |> 
  mutate(CV = round((sd/mean), 2)*100) |> 
  mutate(Time = "Calibration")


#### FOrecasting period 

#cleaned stats for full year for table in MS 
targets |> 
  filter(datetime >= ymd("2024-03-01"),
         datetime <= ymd("2025-04-03")
         ) |> 
  group_by(site_id) |> 
  summarise(min = round(min(observation, na.rm = T), 1),
            median = round(median(observation, na.rm = T), 1),
            mean = round(mean(observation, na.rm = T), 1),
            max = round(max(observation, na.rm = T), 1),
            sd = round(sd(observation, na.rm = T), 1),
            range = max - min,
            range_A = paste(min, max, sep = " - ")) |> 
  mutate(CV_percent = round((sd/mean), 2)*100) |> 
  select(site_id, median, range_A, CV_percent)

#cleaned stats for full year by seasons for table in MS 
targets |> 
     mutate(month = month(datetime),
         Season2 = ifelse(month %in% c(3,4,5), "Spring", NA),
         Season2 = ifelse(month %in% c(6,7,8), "Summer", Season2),
         Season2 = ifelse(month %in% c(9,10,11), "Fall", Season2),
         Season2 = ifelse(month %in% c(1,2,12), "Winter", Season2)
         ) |> 
  filter(datetime >= ymd("2024-03-01"),
         datetime <= ymd("2025-04-03")
         ) |> 
  group_by(site_id, Season2) |> 
  summarise(min = round(min(observation, na.rm = T), 1),
            median = round(median(observation, na.rm = T), 1),
            mean = round(mean(observation, na.rm = T), 1),
            max = round(max(observation, na.rm = T), 1),
            sd = round(sd(observation, na.rm = T), 1),
            range = max - min,
            range_A = paste(min, max, sep = " - ")) |> 
  mutate(CV_percent = round((sd/mean), 2)*100) |> 
  select(site_id, Season2, median, range_A, CV_percent) |> 
  mutate(median = as.character(median),
         CV_percent = as.character(CV_percent)) |> 
  pivot_longer(-c(1:2)) |> 
  pivot_wider(names_from = Season2, values_from = value)



```


water level data and plots 

```{r}
##read in 10 minute data
bvr10min <- read_csv("../Data/Targets/BVR_catwalk_10min.csv")
ccr10min <- read_csv("../Data/Targets/CCR_catwalk_10min.csv")
fcr10min <- read_csv("../Data/Targets/FCR_catwalk_10min.csv")



#### BVR
bvrdepth <- bvr10min |> 
  mutate(Date = as.Date(DateTime)) |> 
  group_by(Date) |> 
  summarise(depth = mean(LvlDepth_m_13, na.rm = T)) |> 
  filter(Date >= ymd("2023-03-13"),
         Date <= ymd("2025-04-03")
         ) |> 
  mutate(site = "BVR")

summary(bvrdepth)

bvrdepth |> 
  ggplot(aes(x = Date, y = depth)) + geom_point()

depth_bvrplot <- bvrdepth |> 
  ggplot(aes(x = Date, y = depth)) + geom_point() + geom_vline(xintercept = ymd("2024-09-26"))+ ggtitle("BVR")
#plotly::ggplotly(depth_bvrplot)
##BVR water level rises from 9.89 to 10.30 (0.4m) between 25sep24 (day before) and 6oct24 (10 days after helene)


#### CCR
ccrdepth <- ccr10min |> 
  mutate(Date = as.Date(DateTime)) |> 
  group_by(Date) |> 
  summarise(depth = mean(LvlDepth_m_13, na.rm = T)) |> 
  filter(Date >= ymd("2023-03-13"),
         Date <= ymd("2025-04-03")) |> 
  mutate(site = "CCR")

summary(ccrdepth)

ccrdepth |> 
  ggplot(aes(x = Date, y = depth)) + geom_point()

# depth_ccrplot <- ccrdepth |> 
#   ggplot(aes(x = Date, y = depth)) + geom_point() + geom_vline(xintercept = ymd("2024-09-26"))+ ggtitle("CCR")
# plotly::ggplotly(depth_ccrplot)
##CCR water level rises from 17.26 to 18.48 (1.2m) between 25sep24 (day before) and 6oct24 (10 days after helene)


#### FCR
fcrdepth <- fcr10min |> 
  mutate(Date = as.Date(DateTime)) |> 
  group_by(Date) |> 
  summarise(depth = mean(LvlDepth_m_9, na.rm = T)) |> 
  filter(Date >= ymd("2024-03-01"),
         Date <= ymd("2025-04-03")) |> 
  mutate(site = "FCR")

summary(fcrdepth)

fcrdepth |> 
  ggplot(aes(x = Date, y = depth)) + geom_point()

# depth_fcrplot <- fcrdepth |> 
#   ggplot(aes(x = Date, y = depth)) + geom_point() + geom_vline(xintercept = ymd("2024-09-26"))+ ggtitle("FCR")
# plotly::ggplotly(depth_fcrplot)
##FCR water level rises from 9.17 to 9.18  (0.01 m) between 25sep24 (day before) and 6oct24 (10 days after helene)
## But peaked at 9.29m on 28sep24 (0.1m)


## combined depth plot
allres_waterlevel <- rbind(fcrdepth, bvrdepth, ccrdepth) |> 
  ggplot(aes(x = Date, y = depth, color = site))+
  geom_point()+
  facet_wrap(~site, ncol = 1, scales = "free_y")+
  scale_x_date(date_breaks = "2 month", date_labels = "%b" )+
  labs(x = "Date", y = "Water level (m)", color = "Reservoir")+
    geom_vline(xintercept = ymd("2024-03-01"), linetype = 2, linewidth = 1.2)+
    #   annotate("text", x = as.Date("2023-08-15"), y = 8, label = "Calibration", 
    #          color = "black", size = 4, fontface = "bold")+
    # annotate("text", x = as.Date("2024-05-15"), y = 8, label = "Forecasting", 
    #          color = "black", size = 4, fontface = "bold")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 18),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = c("FCR" = "#009E73", "BVR" = "orange", "CCR" = "skyblue"))
  
allres_waterlevel

ggsave("../Figures/Fig_S1_waterlevel2.png", allres_waterlevel, height = 5, width = 6.5, units = "in")


plotly::ggplotly(allres_waterlevel)
#from 11feb to 18feb 
#BVR 8.5 - 9.1: 0.6m
#CCR 18.35 - 19.85: 1.5m
#FCR 9.26 to 9.26, but peaked at 9.36 on 16feb

bvrdepth |> 
    ggplot(aes(x = Date, y = depth, color = site))+
 geom_point()+
  facet_wrap(~site, ncol = 1, scales = "free_y")+
  scale_x_date(date_breaks = "2 month", date_labels = "%b" )+
  labs(x = "Date", y = "Water level (m)", color = "Reservoir")+
    geom_vline(xintercept = ymd("2024-03-01"), linetype = 2, linewidth = 1.2)+
      annotate("text", x = as.Date("2023-08-15"), y = 9, label = "Calibration", 
             color = "black", size = 4, fontface = "bold")+
    annotate("text", x = as.Date("2024-05-15"), y = 9, label = "Forecasting", 
             color = "black", size = 4, fontface = "bold")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 18),
                    panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_color_manual(values = c("FCR" = "#009E73", "BVR" = "orange", "CCR" = "skyblue"))


```


## Example forecast plots

```{r}

###set up data for example
weekbefore <- Scored_forecasts |> 
    filter(datetime <= ymd("2024-07-15"),
           datetime >= ymd("2024-07-15") - 7) |> 
  select(datetime, observation, model_id, site_id)

example_forecast_data <- Scored_forecasts |> 
  filter(reference_datetime == ymd("2024-07-15"))
  

### Example for MS 
example_4cast <- example_forecast_data |> 
  plyr::rbind.fill(weekbefore) |> 
  mutate(observation = ifelse(datetime > ymd("2024-07-15"), NA, observation)) |> 
  ggplot()+
  geom_point(aes(datetime, observation)) +
  geom_vline(aes(xintercept = as.Date("2024-07-15")))+
  geom_ribbon(aes(x = datetime, ymin = quantile02.5, ymax = quantile97.5, fill = model_id), alpha = 0.2)+
  geom_line(aes(x = datetime, y = mean, color = model_id))+
  facet_wrap(~site_id, scales = "free", nrow = 3) +
  labs(x = 'Date', y = 'fDOM (QSU)', fill = "model", color = "model") +
  guides(fill = guide_legend(nrow = 6))+
  theme_bw() + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  scale_fill_manual(values = model_colors)+scale_color_manual(values = model_colors)

example_4cast

#ggsave("../Figures/Example_4cast_longer.png", example_4cast, width = 5.5, height = 4.5, units = "in")


#### Examples for workflow figure

#figure where just one model at one reservoir; NO evaluation
example_workflow <-  example_forecast_data |> 
  plyr::rbind.fill(weekbefore) |> 
  filter(site_id == "FCR", model_id == "TS_AR") |> 
  mutate(observation = ifelse(datetime > ymd("2024-07-15"), NA, observation)) |> 
  ggplot()+
  geom_point(aes(datetime, observation)) +
  geom_vline(aes(xintercept = as.Date("2024-07-15")))+
  geom_ribbon(aes(x = datetime, ymin = quantile02.5, ymax = quantile97.5), alpha = 0.2)+
  geom_line(aes(x = datetime, y = mean))+
  labs(x = 'Date', y = 'fDOM (QSU)', fill = "model", color = "model") +
  guides(fill = guide_legend(nrow = 6))+
  theme_bw() + theme(text = element_text(size = 18))+
  scale_fill_manual(values = model_colors)+scale_color_manual(values = model_colors)

example_workflow

#ggsave("../Figures/Ex_4cast_workflow.png", example_workflow, width = 2.8, height = 3, units = "in")



#figure where just one model at one reservoir; YES evaluated
example_workflow_eval <-  example_forecast_data |> 
  plyr::rbind.fill(weekbefore) |>
  filter(site_id == "FCR", model_id == "TS_AR") |> 
  ggplot()+
  geom_point(aes(datetime, observation)) +
  geom_vline(aes(xintercept = as.Date("2024-07-15")))+
  geom_ribbon(aes(x = datetime, ymin = quantile02.5, ymax = quantile97.5), alpha = 0.2)+
  geom_line(aes(x = datetime, y = mean))+
  labs(x = 'Date', y = 'fDOM (QSU)', fill = "model", color = "model") +
  guides(fill = guide_legend(nrow = 6))+
  theme_bw() + theme(text = element_text(size = 18))+
  scale_fill_manual(values = model_colors)+scale_color_manual(values = model_colors)

example_workflow_eval

#ggsave("../Figures/Ex_4cast_workflow_eval.png", example_workflow_eval, width = 2.8, height = 3, units = "in")


```



## Other EXO variable TS

```{r}
#set p for temp corrections 
p <- -0.01

#### set up data
## FCR 
fcr10min <- read_csv("./Data/Targets/FCR_catwalk_10min.csv")

fcr_daily <- fcr10min |> 
  mutate(fdom_TC = EXOfDOM_QSU_1/(1 + (p*(EXOTemp_C_1 - 20)) )   ) |> 
  mutate(Date = as.Date(DateTime)) |> 
  group_by(Date) |> 
  summarise(fDOM_QSU_mean = mean(fdom_TC, na.rm = T),
            Temp_C_mean = mean(EXOTemp_C_1, na.rm = T),
            Chla_ugL_mean = mean(EXOChla_ugL_1, na.rm = T),
            Turb_FNU_mean = mean(EXOTurbidity_FNU_1, na.rm = T),
            Depth_m_mean = mean(LvlDepth_m_9, na.rm = T)
            ) |> 
  mutate(Reservoir = "FCR")


## BVR
bvr10min <- read_csv("./Data/Targets/BVR_catwalk_10min.csv")

bvr_daily <- bvr10min |> 
  mutate(fdom_TC = EXOfDOM_QSU_1.5/(1 + (p*(EXOTemp_C_1.5 - 20)) )   ) |> 
  mutate(Date = as.Date(DateTime)) |> 
  group_by(Date) |> 
  summarise(fDOM_QSU_mean = mean(fdom_TC, na.rm = T),
            Temp_C_mean = mean(EXOTemp_C_1.5, na.rm = T),
            Chla_ugL_mean = mean(EXOChla_ugL_1.5, na.rm = T),
            Turb_FNU_mean = mean(EXOTurbidity_FNU_1.5, na.rm = T),
            Depth_m_mean = mean(LvlDepth_m_13, na.rm = T)
  ) |> 
  mutate(Reservoir = "BVR")


## CCR
ccr10min <- read_csv("./Data/Targets/CCR_catwalk_10min.csv")

ccr_daily <- ccr10min |> 
  mutate(fdom_TC = EXOfDOM_QSU_1/(1 + (p*(EXOTemp_C_1 - 20)) )   ) |> 
  mutate(Date = as.Date(DateTime)) |> 
  group_by(Date) |> 
  summarise(fDOM_QSU_mean = mean(fdom_TC, na.rm = T),
            Temp_C_mean = mean(EXOTemp_C_1, na.rm = T),
            Chla_ugL_mean = mean(EXOChla_ugL_1, na.rm = T),
            #Turb_FNU_mean = mean(EXOTurbidity_FNU_1, na.rm = T),
            Depth_m_mean = mean(LvlDepth_m_13, na.rm = T)
  ) |> 
  mutate(Reservoir = "CCR")


### bind data
daily_data <- plyr::rbind.fill(fcr_daily, ccr_daily, bvr_daily)


#### Plots 

#sensors
exo_othervariables <- daily_data |> 
  filter(Date >= ymd("2023-03-13"),
         Date <= ymd("2025-04-03")) |> 
  rename(Temp_C = Temp_C_mean,
         Chla_ugL = Chla_ugL_mean,
         Turb_FNU = Turb_FNU_mean,
         fDOM_QSU = fDOM_QSU_mean) |> 
  select(-Depth_m_mean) |> 
  pivot_longer(-c(Reservoir, Date)) |> 
  ggplot(aes(x = Date, y = value))+
  geom_point()+
  geom_vline(xintercept = ymd("2024-03-01"), linetype = 2, linewidth = 1.2)+
  facet_grid(factor(name, levels = c("Temp_C", "fDOM_QSU", "Turb_FNU", "Chla_ugL"))~Reservoir,
             scales = "free_y")+
  scale_x_date(date_labels = "%b", breaks = "6 months")+
  labs(x = "Date", y = "Variable", color = "Reservoir")+
  theme_bw() + theme(legend.position = "top", text = element_text(size = 14),
                     panel.grid.major = element_blank(), panel.grid.minor = element_blank())


#ggsave("../Figures/SI_exo_timeseries.png", exo_othervariables, width = 6, height = 4.5, units = "in")

#R2R Figure
daily_data |>
  filter(Date >= ymd("2023-03-13"),
         Date <= ymd("2025-04-03")) |> 
  filter(Reservoir != "CCR") |> 
  ggplot(aes(x = Turb_FNU_mean, y = fDOM_QSU_mean))+
  geom_point()+
  geom_smooth(method = "lm", se = FALSE, color = "blue") +
  stat_poly_eq(
    aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")),
    formula = y ~ x,
    parse = TRUE,
    size = 3
  )+
  facet_wrap(~Reservoir, scales = "free")
  




```
